+++ cypress/e2e/pre-release-tests/getOpenAdministrations.cy.js
@@ -1,6 +1,6 @@
 import { signInAsSuperAdmin } from '../../support/helper-functions/super-admin/superAdminHelpers';
-import { getDevFirebase } from '../../support/helper-functions/devFirebase';
-import { getOpenAdministrations } from '../../support/helper-functions/query';
+import { getDevFirebase } from '../../support/devFirebase';
+import { getOpenAdministrations } from '../../support/query';
 
 const adminAuth = getDevFirebase('admin').auth;
 const adminFirestore = getDevFirebase('admin').db;
